+++ linearmodels/asset_pricing/results.py
@@ -6,11 +6,11 @@ from __future__ import annotations
 from linearmodels.compat.statsmodels import Summary
 
 import datetime as dt
+from functools import cached_property
 from typing import cast
 
 import numpy as np
 import pandas as pd
-from property_cached import cached_property
 from scipy import stats
 from statsmodels.iolib.summary import SimpleTable, fmt_2cols, fmt_params