+++ linearmodels/panel/results.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 from linearmodels.compat.statsmodels import Summary
 
 import datetime as dt
+from functools import cached_property
 from typing import Any, Mapping, Union
 
 from formulaic.utils.context import capture_context
 import numpy as np
 from pandas import DataFrame, Series, concat
-from property_cached import cached_property
 from scipy import stats
 from statsmodels.iolib.summary import SimpleTable, fmt_2cols, fmt_params